Split GoldensClient out of flutter_goldens.dart
This enables using GoldensClient from the flutter driver file for the
embedded Android views test.

flutter_goldens.dart has a transitive dependency on dart:ui so it cannot
be included in non Flutter apps. Splitting it out allows us to use it
from driver tests.

This also removes the visibleForTesting annotation from GoldensClient.
